# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Real Estate Helper API"
  description: "This is an API to help access United States real estate county data and compare patterns over the years"
  version: "1.0.0"
host: "127.0.0.1:5000" #"my-real-estate-helper.onrender.com"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header



paths: 
  /users/login:
    post:
      tags:
        - Users
      summary: "Login endpoint and generates token"
      description: "Validates User credentials (email and password), and creates a user specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema:
            $ref: "#/definitions/LoginSuccessResponse"
          examples:
            application/json:
              message: "Welcome User"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOj..."
              user: 
                id: 0
                first_name: "New"
                last_name: "Guy"
                email: "newguy@email.com"
                password: "<encrypted password>"
                role: "free_user"

        401:
          description: "Login Failed"
          schema:
            $ref: "#/definitions/LoginFailResponse"
          examples:
            application/json:
              message: "Failed to find User credentials"
  
  /users:
    post:
      tags:
        - Users
      summary: "Create User"
      description: "Send in User info, to create a User profile that will be stored in the database."
      parameters:
        - in: "body"
          name: "body"
          description: "New User information"
          required: true
          schema:
            $ref: "#/definitions/UserInput"

      responses:
        201:
          description: "User creation Successful"
          schema:
            $ref: "#/definitions/UserResponse"
          examples:
            application/json:
              id: 0
              first_name: "New"
              last_name: "Guy"
              email: "newguy@email.com"
              password: "<encrypted password>"
              role: "free_user"

    get:
      tags:
        - Users
      summary: "Gets a User Profile"
      description: "View User's profile. Requires a token. User will be identified by token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "User View Profile"
          required: true
          schema:
            $ref: "#/definitions/UserInput"
      
      responses:
        200:
          description: "Successfully retrieved User profile."
          schema:
            $ref: "#/definitions/UserResponse"
          examples:
            application/json:
              id: 0
              first_name: "New"
              last_name: "Guy"
              email: "newguy@email.com"
              password: "<encrypted password>"
              role: "free_user"
              
    put:
      tags:
        - Users
      summary: "Updates a User"
      description: "Endpoint to update a User profile. Requires a token. User will be identified by token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "User Updates"
          required: true
          schema:
            $ref: "#/definitions/UserInput"
      
      responses:
        200:
          description: "Successfully updated User profile."
          schema:
            $ref: "#/definitions/UserResponse"
          examples:
            application/json:
              id: 0
              first_name: "New"
              last_name: "Guy"
              email: "newguy@email.com"
              password: "<encrypted password>"
              role: "free_user"

    delete:
      tags:
        - Users
      summary: "Deletes a User"
      description: "Deletes a User object. Requires a token. Token will be used to identify User to delete."
      security:
        - bearerAuth: []

      responses:
        200:
          description: "Successfully deleted User profile."
          schema:
            type: object
            properties:
              message:
                type: string


  /users/view_compare_counties:
    get:
      tags:
        - Compare_county
      summary: "Returns all counties assigned to User"
      description: "Endpoint to receive an array of counties assigned to User. Token is required to identify User."
      responses:
        200:
          description: "Successfully Retrieved User's compare counties list"
          schema:
            $ref: "#/definitions/CountyNameMappingResponse"

  /users/assign_compare_county/{county_fips}:
      post:
        tags:
          - Compare_county
        summary: "Adds a Compare_county to User's county compare list"
        description: "Endpoint to add a Compare_county object assigned to User. Token is required to identify User."
        parameters:
        - in: path
          name: county_fips
          description: "New compare_county information"
          required: true
          schema:
            $ref: "#/definitions/Compare_CountyInput"
        responses:
          200:
            description: "Successfully added a Compare_county"
            schema:
              type: object
              properties:
                message:
                  type: string

  /users/remove_compare_county/{county_fips}:
      delete:
        tags:
          - Compare_county
        summary: "Removes a Compare_county from User's county compare list"
        description: "Removes a Compare_county object. Token is required to identify User. Requires a county_fips id to identify which compare_county to remove."
        parameters:
          - in: path
            name: county_fips
            required: true
            type: string
            description: Fips ID of the compare_county to delete
        responses:
          200:
            description: "Successfully removed a Compare_county."
            schema:
              type: object
              properties:
                message:
                  type: string



  /county_data/upload:
    post:
      tags:
        - County_data
      summary: "Creates the historical county_data by uploading csv file"
      description: "Creates county_data objects based on the data from the historical csv file."
      parameters:
        - in: formData
          name: file
          type: file
          description: "RDC_Inventory_Core_Metrics_County_History.csv"
          required: true
      
      responses:
        200:
          description: "Successfully created county_data"
          schema:
            type: object
            properties:
              message:
                type: string

  
  /county_data/{fips_id}:
    get:
      tags:
        - County_data
      summary: "Gets all county_data by using fips_id"
      description: "Gets an array of county_data objects based on the fips_id."
      parameters:
        - in: path
          name: fips_id
          description: "County fips id"
          required: true
          
      responses:
        200:
          description: "Successfully Retrieved all historical county_data for county matching the fips id."
          schema:
            $ref: "#/definitions/CountyDataResponse"


  /county_name_mapping/upload:
    post:
      tags:
        - County_name_mapping
      summary: "Creates the county_name_mapping by uploading csv file"
      description: "Creates county_name_mapping objects based on the data from a csv file."
      parameters:
        - in: formData
          name: file
          type: file
          description: "RDC_Inventory_Core_Metrics_County.csv"
          required: true
      
      responses:
        200:
          description: "Successfully created county_name_mapping"
          schema:
            type: object
            properties:
              message:
                type: string

  /county_name_mapping:
    get:
      tags:
        - County_name_mapping
      summary: "Gets all current county names in the United States"
      description: "Gets an array of county_name_mapping objects. This matches a fips_id with the county name."
          
      responses:
        200:
          description: "Successfully Retrieved all current county_name_mapping objects."
          schema:
            $ref: "#/definitions/CountyNameMappingResponse"



definitions: 

#-------------------User Login---------------
  LoginCredentials: #Input data required for login route
    type: object
    properties:
      email:
        type: string
      password:
        type: string

    required:
      - email
      - password

  LoginSuccessResponse: #Response from login endpoint
    type: object
    properties:
      message:
        type: string
      token:
        type: string
      mechanic:
       $ref: '#/definitions/UserResponse'

  LoginFailResponse:
    type: object
    properties:
      message:
        type: string

# -------------------

#-------------------User Input---------------

  UserInput: #Sending user Information
    type: object
    properties:
      first_name:
        type: string
        example: "New"
      last_name:
        type: string
        example: "Guy"
      email:
        type: string
        example: "newguy@email.com"
      password:
        type: string
        example: "bestguy"
      role:
        type: string
        example: "free_user" 

      
    required:
      - first_name
      - last_name
      - email
      - password
    
  UserResponse: 
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        type: string


#--------------------



#-------------------County_data Response---------------

  CountyDataResponse:
    type: array
    items: 
      type: object
      properties:
        info_date:
          type: string
          format: date
          example: 2025-01-01
        fips_id:
          type: string
          example: "12345"
        median_listing_price:
          type: float
          example: 300500.0
        active_listing_count:
          type: integer
          example: 123
        active_listing_count_yy:
          type: float
          example: 0.123
        median_days_on_market:
          type: integer
          example: 123
        price_reduced_count:
          type: integer
          example: 12
        pending_listing_count:
          type: integer
          example: 123
 
# -------------------



#-------------------County Name Mapping Response---------------

  CountyNameMappingResponse:
    type: array
    items: 
      type: object
      properties:
        fips_id:
          type: string
          example: "12345"
        county_name:
          type: string
          example: "county_name"
       
# -------------------